// packages/core/src/API.ts
export type Place_orderMutation = {
  place_order?: string | null;
};

export type Cancel_orderMutationVariables = {
  input: UserActionInput;
};

export type Cancel_orderMutation = {
  cancel_order?: string | null;
};

export type Cancel_allMutationVariables = {
  input: UserActionInput;
};

export type Cancel_allMutation = {
  cancel_all?: string | null;
};

export type WithdrawMutationVariables = {
  input: UserActionInput;
};

export type WithdrawMutation = {
  withdraw?: string | null;
};